requirements.txt
beautifulsoup4
jupyter
keras
matplotlib
numpy
opencv-python
pandas
pillow
pydot
scikit-learn
scipy
seaborn
tensorflow